Better refactoring
